+++ src/helpers.php
@@ -57,7 +57,7 @@ if (!function_exists('generate_sentence_from_array')) {
     function generate_sentence_from_array(
         array $stringParts,
         string $delimiter = ', ',
-        string $lastDelimiter = ' and '
+        string $lastDelimiter = ', and '
     ): string {
         return str_replace_last($delimiter, $lastDelimiter, implode($delimiter, $stringParts));
     }